+++ packages/core/js/framework7-lazy.js
@@ -1,5 +1,5 @@
 /**
- * Framework7 3.4.3
+ * Framework7 3.5.0
  * Full featured mobile HTML framework for building iOS & Android apps
  * http://framework7.io/
  *
@@ -7,7 +7,7 @@
  *
  * Released under the MIT License
  *
- * Released on: October 19, 2018
+ * Released on: October 26, 2018
  */
 
 (function (global, factory) {
@@ -5961,7 +5961,8 @@
     }
     $newPage
       .addClass(("page-" + newPagePosition))
-      .removeClass('stacked');
+      .removeClass('stacked')
+      .trigger('page:unstack');
 
     if (dynamicNavbar && $newNavbarInner.length) {
       $newNavbarInner
@@ -5994,6 +5995,7 @@
           var oldNavbarInnerEl = app.navbar.getElByPage($pagesInView.eq(i));
           if (router.params.stackPages) {
             $pagesInView.eq(i).addClass('stacked');
+            $pagesInView.eq(i).trigger('page:stack');
             if (separateNavbar) {
               // $navbarsInView.eq(i).addClass('stacked');
               $(oldNavbarInnerEl).addClass('stacked');
@@ -6108,6 +6110,7 @@
     if (options.reloadCurrent && $oldPage.length > 0) {
       if (router.params.stackPages && router.initialPages.indexOf($oldPage[0]) >= 0) {
         $oldPage.addClass('stacked');
+        $oldPage.trigger('page:stack');
         if (separateNavbar) {
           $oldNavbarInner.addClass('stacked');
         }
@@ -6125,6 +6128,7 @@
         var $oldNavbarInnerEl = $(app.navbar.getElByPage($oldPageEl));
         if (router.params.stackPages && router.initialPages.indexOf($oldPageEl[0]) >= 0) {
           $oldPageEl.addClass('stacked');
+          $oldPageEl.trigger('page:stack');
           if (separateNavbar) {
             $oldNavbarInnerEl.addClass('stacked');
           }
@@ -6140,6 +6144,7 @@
     } else if (options.reloadPrevious) {
       if (router.params.stackPages && router.initialPages.indexOf($oldPage[0]) >= 0) {
         $oldPage.addClass('stacked');
+        $oldPage.trigger('page:stack');
         if (separateNavbar) {
           $oldNavbarInner.addClass('stacked');
         }
@@ -6204,6 +6209,7 @@
       if (!keepOldPage) {
         if (router.params.stackPages) {
           $oldPage.addClass('stacked');
+          $oldPage.trigger('page:stack');
           if (separateNavbar) {
             $oldNavbarInner.addClass('stacked');
           }
@@ -7009,7 +7015,8 @@
     $newPage
       .addClass('page-previous')
       .removeClass('stacked')
-      .removeAttr('aria-hidden');
+      .removeAttr('aria-hidden')
+      .trigger('page:unstack');
 
     if (dynamicNavbar && $newNavbarInner.length > 0) {
       $newNavbarInner
@@ -7044,6 +7051,7 @@
             if ($pageToRemove[0] !== $newPage[0] && $pageToRemove.index() > $newPage.index()) {
               if (router.initialPages.indexOf($pageToRemove[0]) >= 0) {
                 $pageToRemove.addClass('stacked');
+                $pageToRemove.trigger('page:stack');
                 if (separateNavbar) {
                   $navbarToRemove.addClass('stacked');
                 }
@@ -7065,6 +7073,7 @@
           }
           if (router.params.stackPages && router.initialPages.indexOf($pageToRemove[0]) >= 0) {
             $pageToRemove.addClass('stacked');
+            $pageToRemove.trigger('page:stack');
             $navbarToRemove.addClass('stacked');
           } else if ($pageToRemove.length > 0) {
             router.pageCallback('beforeRemove', $pageToRemove, $navbarToRemove, 'previous', undefined, options);
@@ -7130,6 +7139,7 @@
           }
           if (router.params.stackPages && router.initialPages.indexOf(pageToRemove) >= 0) {
             $pageToRemove.addClass('stacked');
+            $pageToRemove.trigger('page:stack');
             if (separateNavbar) {
               $navbarToRemove.addClass('stacked');
             }
@@ -7217,6 +7227,7 @@
       // Remove Old Page
       if (router.params.stackPages && router.initialPages.indexOf($oldPage[0]) >= 0) {
         $oldPage.addClass('stacked');
+        $oldPage.trigger('page:stack');
         if (separateNavbar) {
           $oldNavbarInner.addClass('stacked');
         }