+++ packages/react/components/link.d.ts
@@ -47,7 +47,6 @@ declare namespace F7Link {
     force? : boolean
     animate? : boolean
     ignoreCache? : boolean
-    pageName? : string
     reloadCurrent? : boolean
     reloadAll? : boolean
     reloadPrevious? : boolean