+++ packages/react/components/list-button.d.ts
@@ -27,7 +27,6 @@ declare namespace F7ListButton {
     force? : boolean
     animate? : boolean
     ignoreCache? : boolean
-    pageName? : string
     reloadCurrent? : boolean
     reloadAll? : boolean
     reloadPrevious? : boolean