+++ packages/react/components/page.js
@@ -14,7 +14,8 @@ class F7Page extends React.Component {
     this.state = (() => {
       return {
         hasSubnavbar: false,
-        routerClasses: ''
+        routerClass: '',
+        routerForceUnstack: false
       };
     })();
   }
@@ -49,6 +50,18 @@ class F7Page extends React.Component {
     this.dispatchEvent('page:mounted pageMounted', event, page);
   }
 
+  onPageStack() {
+    this.setState({
+      routerForceUnstack: false
+    });
+  }
+
+  onPageUnstack() {
+    this.setState({
+      routerForceUnstack: true
+    });
+  }
+
   onPageInit(event) {
     const page = event.detail;
     const {
@@ -77,13 +90,13 @@ class F7Page extends React.Component {
 
     if (page.from === 'next') {
       this.setState({
-        routerClasses: 'page-next'
+        routerClass: 'page-next'
       });
     }
 
     if (page.from === 'previous') {
       this.setState({
-        routerClasses: 'page-previous'
+        routerClass: 'page-previous'
       });
     }
 
@@ -100,13 +113,13 @@ class F7Page extends React.Component {
 
     if (page.to === 'next') {
       this.setState({
-        routerClasses: 'page-next'
+        routerClass: 'page-next'
       });
     }
 
     if (page.to === 'previous') {
       this.setState({
-        routerClasses: 'page-previous'
+        routerClass: 'page-previous'
       });
     }
 
@@ -116,7 +129,7 @@ class F7Page extends React.Component {
   onPageAfterIn(event) {
     const page = event.detail;
     this.setState({
-      routerClasses: 'page-current'
+      routerClass: 'page-current'
     });
     this.dispatchEvent('page:afterin pageAfterIn', event, page);
   }
@@ -196,8 +209,8 @@ class F7Page extends React.Component {
     }
 
     const forceSubnavbar = typeof subnavbar === 'undefined' && typeof withSubnavbar === 'undefined' ? hasSubnavbar || this.state.hasSubnavbar : false;
-    const classes = Utils.classNames(className, 'page', this.state.routerClasses, {
-      stacked,
+    const classes = Utils.classNames(className, 'page', this.state.routerClass, {
+      stacked: stacked && !this.state.routerForceUnstack,
       tabs,
       'page-with-subnavbar': subnavbar || withSubnavbar || forceSubnavbar,
       'no-navbar': noNavbar,
@@ -259,6 +272,8 @@ class F7Page extends React.Component {
     el.removeEventListener('page:afterout', self.onPageAfterOut);
     el.removeEventListener('page:afterin', self.onPageAfterIn);
     el.removeEventListener('page:beforeremove', self.onPageBeforeRemove);
+    el.removeEventListener('page:stack', self.onPageStack);
+    el.removeEventListener('page:unstack', self.onPageUnstack);
   }
 
   componentDidMount() {
@@ -282,6 +297,8 @@ class F7Page extends React.Component {
     self.onPageAfterOut = self.onPageAfterOut.bind(self);
     self.onPageAfterIn = self.onPageAfterIn.bind(self);
     self.onPageBeforeRemove = self.onPageBeforeRemove.bind(self);
+    self.onPageStack = self.onPageStack.bind(self);
+    self.onPageUnstack = self.onPageUnstack.bind(self);
 
     if (ptr) {
       el.addEventListener('ptr:pullstart', self.onPtrPullStart);
@@ -303,6 +320,8 @@ class F7Page extends React.Component {
     el.addEventListener('page:afterout', self.onPageAfterOut);
     el.addEventListener('page:afterin', self.onPageAfterIn);
     el.addEventListener('page:beforeremove', self.onPageBeforeRemove);
+    el.addEventListener('page:stack', self.onPageStack);
+    el.addEventListener('page:unstack', self.onPageUnstack);
   }
 
   get slots() {