+++ packages/react/framework7-react.js
@@ -1,5 +1,5 @@
 /**
- * Framework7 React 3.4.3
+ * Framework7 React 3.5.0
  * Build full featured iOS & Android apps using Framework7 & React
  * http://framework7.io/react/
  *
@@ -7,7 +7,7 @@
  *
  * Released under the MIT License
  *
- * Released on: October 19, 2018
+ * Released on: October 26, 2018
  */
 
 (function (global, factory) {
@@ -153,7 +153,6 @@
         default: undefined,
       },
       ignoreCache: Boolean,
-      pageName: String,
       reloadCurrent: Boolean,
       reloadAll: Boolean,
       reloadPrevious: Boolean,
@@ -3300,6 +3299,7 @@
         var defaultValue = props.defaultValue;
         return {
           inputFocused: false,
+          inputInvalid: false,
           currentInputValue: typeof value === 'undefined' ? defaultValue : value
         };
       })();
@@ -3323,6 +3323,26 @@
 
     var prototypeAccessors = { inputWithValue: { configurable: true },slots: { configurable: true },refs: { configurable: true } };
 
+    F7Input.prototype.validateInput = function validateInput (inputEl) {
+      var self = this;
+      var f7 = self.$f7;
+      if (!f7 || !inputEl) { return; }
+      var validity = inputEl.validity;
+      if (!validity) { return; }
+
+      if (!validity.valid) {
+        if (self.state.inputInvalid !== true) {
+          self.setState({
+            inputInvalid: true
+          });
+        }
+      } else if (self.state.inputInvalid !== false) {
+        self.setState({
+          inputInvalid: false
+        });
+      }
+    };
+
     F7Input.prototype.onTextareaResize = function onTextareaResize (event) {
       this.dispatchEvent('textarea:resize textareaResize', event);
     };
@@ -3340,8 +3360,16 @@
     };
 
     F7Input.prototype.onInput = function onInput (event) {
-      this.dispatchEvent('input', event);
-      this.setState({
+      var self = this;
+      var ref = self.props;
+      var validate = ref.validate;
+      self.dispatchEvent('input', event);
+
+      if ((validate || validate === '') && self.refs && self.refs.inputEl) {
+        self.validateInput(self.refs.inputEl);
+      }
+
+      self.setState({
         currentInputValue: event.target.value
       });
     };
@@ -3354,8 +3382,16 @@
     };
 
     F7Input.prototype.onBlur = function onBlur (event) {
-      this.dispatchEvent('blur', event);
-      this.setState({
+      var self = this;
+      var ref = self.props;
+      var validate = ref.validate;
+      self.dispatchEvent('blur', event);
+
+      if ((validate || validate === '') && self.refs && self.refs.inputEl) {
+        self.validateInput(self.refs.inputEl);
+      }
+
+      self.setState({
         inputFocused: false
       });
     };
@@ -3418,7 +3454,7 @@
         var inputClassName = Utils.classNames(!wrap && className, {
           resizable: type === 'textarea' && resizable,
           'no-store-data': noFormStoreData || noStoreData || ignoreStoreData,
-          'input-invalid': errorMessage && errorMessageForce,
+          'input-invalid': errorMessage && errorMessageForce || self.state.inputInvalid,
           'input-with-value': self.inputWithValue,
           'input-focused': self.state.inputFocused
         });
@@ -3533,10 +3569,9 @@
       var self = this;
       var ref = self.props;
       var value = ref.value;
-      var defaultValue = ref.defaultValue;
       var ref$1 = self.state;
       var currentInputValue = ref$1.currentInputValue;
-      return typeof value === 'undefined' ? defaultValue || defaultValue === 0 || currentInputValue : value || value === 0;
+      return typeof value === 'undefined' ? currentInputValue : value || value === 0;
     };
 
     F7Input.prototype.componentWillUnmount = function componentWillUnmount () {
@@ -3590,7 +3625,7 @@
         f7.input.checkEmptyState(inputEl);
 
         if (validate) {
-          f7.input.validate(inputEl);
+          self.validateInput(inputEl);
         }
 
         if (resizable) {
@@ -3627,7 +3662,7 @@
 
         if ((validate || validate === '') && (typeof value !== 'undefined' && value !== null && value !== '' || typeof defaultValue !== 'undefined' && defaultValue !== null && defaultValue !== '')) {
           setTimeout(function () {
-            f7.input.validate(inputEl);
+            self.validateInput(inputEl);
           }, 0);
         }
 
@@ -4290,6 +4325,470 @@
 
   F7ListIndex.displayName = 'f7-list-index';
 
+  var F7ListInput = (function (superclass) {
+    function F7ListInput(props, context) {
+      var this$1 = this;
+
+      superclass.call(this, props, context);
+      this.__reactRefs = {};
+
+      this.state = (function () {
+        var value = props.value;
+        var defaultValue = props.defaultValue;
+        return {
+          isSortable: props.sortable,
+          inputFocused: false,
+          inputInvalid: false,
+          currentInputValue: typeof value === 'undefined' ? defaultValue : value
+        };
+      })();
+
+      (function () {
+        var self = this$1;
+        self.onChangeBound = self.onChange.bind(self);
+        self.onInputBound = self.onInput.bind(self);
+        self.onFocusBound = self.onFocus.bind(self);
+        self.onBlurBound = self.onBlur.bind(self);
+        self.onTextareaResizeBound = self.onTextareaResize.bind(self);
+        self.onInputNotEmptyBound = self.onInputNotEmpty.bind(self);
+        self.onInputEmptyBound = self.onInputEmpty.bind(self);
+        self.onInputClearBound = self.onInputClear.bind(self);
+      })();
+    }
+
+    if ( superclass ) F7ListInput.__proto__ = superclass;
+    F7ListInput.prototype = Object.create( superclass && superclass.prototype );
+    F7ListInput.prototype.constructor = F7ListInput;
+
+    var prototypeAccessors = { inputHasValue: { configurable: true },slots: { configurable: true },refs: { configurable: true } };
+
+    F7ListInput.prototype.validateInput = function validateInput (inputEl) {
+      var self = this;
+      var f7 = self.$f7;
+      if (!f7 || !inputEl) { return; }
+      var validity = inputEl.validity;
+      if (!validity) { return; }
+
+      if (!validity.valid) {
+        if (self.state.inputInvalid !== true) {
+          self.setState({
+            inputInvalid: true
+          });
+        }
+      } else if (self.state.inputInvalid !== false) {
+        self.setState({
+          inputInvalid: false
+        });
+      }
+    };
+
+    F7ListInput.prototype.onTextareaResize = function onTextareaResize (event) {
+      this.dispatchEvent('textarea:resize textareaResize', event);
+    };
+
+    F7ListInput.prototype.onInputNotEmpty = function onInputNotEmpty (event) {
+      this.dispatchEvent('input:notempty inputNotEmpty', event);
+    };
+
+    F7ListInput.prototype.onInputEmpty = function onInputEmpty (event) {
+      this.dispatchEvent('input:empty inputEmpty', event);
+    };
+
+    F7ListInput.prototype.onInputClear = function onInputClear (event) {
+      this.dispatchEvent('input:clear inputClear', event);
+    };
+
+    F7ListInput.prototype.onInput = function onInput (event) {
+      var self = this;
+      var ref = self.props;
+      var validate = ref.validate;
+      self.dispatchEvent('input', event);
+
+      if ((validate || validate === '') && self.refs && self.refs.inputEl) {
+        self.validateInput(self.refs.inputEl);
+      }
+
+      self.setState({
+        currentInputValue: event.target.value
+      });
+    };
+
+    F7ListInput.prototype.onFocus = function onFocus (event) {
+      this.dispatchEvent('focus', event);
+      this.setState({
+        inputFocused: true
+      });
+    };
+
+    F7ListInput.prototype.onBlur = function onBlur (event) {
+      var self = this;
+      var ref = self.props;
+      var validate = ref.validate;
+      self.dispatchEvent('blur', event);
+
+      if ((validate || validate === '') && self.refs && self.refs.inputEl) {
+        self.validateInput(self.refs.inputEl);
+      }
+
+      self.setState({
+        inputFocused: false
+      });
+    };
+
+    F7ListInput.prototype.onChange = function onChange (event) {
+      this.dispatchEvent('change', event);
+    };
+
+    prototypeAccessors.inputHasValue.get = function () {
+      var self = this;
+      var ref = self.props;
+      var value = ref.value;
+      var ref$1 = self.state;
+      var currentInputValue = ref$1.currentInputValue;
+      return typeof value === 'undefined' ? currentInputValue : value || value === 0;
+    };
+
+    F7ListInput.prototype.render = function render () {
+      var this$1 = this;
+
+      var self = this;
+      var ref = self.state;
+      var inputFocused = ref.inputFocused;
+      var inputInvalid = ref.inputInvalid;
+      var props = self.props;
+      var id = props.id;
+      var style = props.style;
+      var className = props.className;
+      var sortable = props.sortable;
+      var media = props.media;
+      var renderInput = props.input;
+      var type = props.type;
+      var name = props.name;
+      var value = props.value;
+      var defaultValue = props.defaultValue;
+      var readonly = props.readonly;
+      var required = props.required;
+      var disabled = props.disabled;
+      var placeholder = props.placeholder;
+      var inputId = props.inputId;
+      var size = props.size;
+      var accept = props.accept;
+      var autocomplete = props.autocomplete;
+      var autocorrect = props.autocorrect;
+      var autocapitalize = props.autocapitalize;
+      var spellcheck = props.spellcheck;
+      var autofocus = props.autofocus;
+      var autosave = props.autosave;
+      var max = props.max;
+      var min = props.min;
+      var step = props.step;
+      var maxlength = props.maxlength;
+      var minlength = props.minlength;
+      var multiple = props.multiple;
+      var inputStyle = props.inputStyle;
+      var pattern = props.pattern;
+      var validate = props.validate;
+      var tabindex = props.tabindex;
+      var resizable = props.resizable;
+      var clearButton = props.clearButton;
+      var noFormStoreData = props.noFormStoreData;
+      var noStoreData = props.noStoreData;
+      var ignoreStoreData = props.ignoreStoreData;
+      var errorMessage = props.errorMessage;
+      var errorMessageForce = props.errorMessageForce;
+      var info = props.info;
+      var label = props.label;
+      var inlineLabel = props.inlineLabel;
+      var floatingLabel = props.floatingLabel;
+      var isSortable = sortable || self.state.isSortable;
+
+      var createInput = function (tag, children) {
+        var InputTag = tag;
+        var needsValue = type !== 'file';
+        var needsType = tag === 'input';
+        var inputClassName = Utils.classNames({
+          resizable: type === 'textarea' && resizable,
+          'no-store-data': noFormStoreData || noStoreData || ignoreStoreData,
+          'input-invalid': errorMessage && errorMessageForce || inputInvalid,
+          'input-with-value': self.inputHasValue,
+          'input-focused': inputFocused
+        });
+        var input;
+        {
+          input = React.createElement(InputTag, {
+            ref: function (__reactNode) {
+              this$1.__reactRefs['inputEl'] = __reactNode;
+            },
+            style: inputStyle,
+            name: name,
+            type: needsType ? type : undefined,
+            placeholder: placeholder,
+            id: inputId,
+            value: needsValue ? value : undefined,
+            defaultValue: defaultValue,
+            size: size,
+            accept: accept,
+            autoComplete: autocomplete,
+            autoCorrect: autocorrect,
+            autoCapitalize: autocapitalize,
+            spellCheck: spellcheck,
+            autoFocus: autofocus,
+            autoSave: autosave,
+            disabled: disabled,
+            max: max,
+            maxLength: maxlength,
+            min: min,
+            minLength: minlength,
+            step: step,
+            multiple: multiple,
+            readOnly: readonly,
+            required: required,
+            pattern: pattern,
+            validate: typeof validate === 'string' && validate.length ? validate : undefined,
+            'data-validate': validate === true || validate === '' ? true : undefined,
+            tabIndex: tabindex,
+            'data-error-message': errorMessageForce ? undefined : errorMessage,
+            className: inputClassName,
+            onFocus: self.onFocusBound,
+            onBlur: self.onBlurBound,
+            onInput: self.onInputBound,
+            onChange: self.onChangeBound
+          }, children);
+        }
+        return input;
+      };
+
+      var inputEl;
+
+      if (renderInput) {
+        if (type === 'select' || type === 'textarea' || type === 'file') {
+          if (type === 'select') {
+            inputEl = createInput('select', self.slots.default);
+          } else if (type === 'file') {
+            inputEl = createInput('input');
+          } else {
+            inputEl = createInput('textarea');
+          }
+        } else {
+          inputEl = createInput('input');
+        }
+      }
+
+      var hasErrorMessage = !!errorMessage || self.slots['error-message'] && self.slots['error-message'].length;
+      return React.createElement('li', {
+        ref: function (__reactNode) {
+          this$1.__reactRefs['el'] = __reactNode;
+        },
+        id: id,
+        style: style,
+        className: Utils.classNames(className, {
+          disabled: disabled
+        }, Mixins.colorClasses(props))
+      }, this.slots['root-start'], React.createElement('div', {
+        className: Utils.classNames('item-content item-input', {
+          'inline-label': inlineLabel,
+          'item-input-focused': inputFocused,
+          'item-input-with-info': !!info || self.slots.info && self.slots.info.length,
+          'item-input-with-value': self.inputHasValue,
+          'item-input-with-error-message': hasErrorMessage && errorMessageForce || inputInvalid,
+          'item-input-invalid': hasErrorMessage && errorMessageForce || inputInvalid
+        })
+      }, this.slots['content-start'], (media || self.slots.media) && React.createElement('div', {
+        className: 'item-media'
+      }, media && React.createElement('img', {
+        src: media
+      }), this.slots['media']), React.createElement('div', {
+        className: 'item-inner'
+      }, this.slots['inner-start'], (label || self.slots.label) && React.createElement('div', {
+        className: Utils.classNames('item-title item-label', {
+          'item-floating-label': floatingLabel
+        })
+      }, label, this.slots['label']), React.createElement('div', {
+        className: Utils.classNames('item-input-wrap', {
+          'input-dropdown': type === 'select'
+        })
+      }, inputEl, this.slots['input'], hasErrorMessage && errorMessageForce && React.createElement('div', {
+        className: 'item-input-error-message'
+      }, errorMessage, this.slots['error-message']), clearButton && React.createElement('span', {
+        className: 'input-clear-button'
+      }), (info || self.slots.info) && React.createElement('div', {
+        className: 'item-input-info'
+      }, info, this.slots['info'])), this.slots['inner'], this.slots['inner-end']), this.slots['content'], this.slots['content-end']), isSortable && React.createElement('div', {
+        className: 'sortable-handler'
+      }), this.slots['root'], this.slots['root-end']);
+    };
+
+    F7ListInput.prototype.componentWillUnmount = function componentWillUnmount () {
+      var self = this;
+      var inputEl = self.refs.inputEl;
+      if (!inputEl) { return; }
+      inputEl.removeEventListener('input:notempty', self.onInputNotEmptyBound, false);
+      inputEl.removeEventListener('textarea:resze', self.onTextareaResizeBound, false);
+      inputEl.removeEventListener('input:empty', self.onInputEmptyBound, false);
+      inputEl.removeEventListener('input:clear', self.onInputClearBound, false);
+    };
+
+    F7ListInput.prototype.componentDidUpdate = function componentDidUpdate (prevProps, prevState) {
+      var this$1 = this;
+
+      __reactComponentWatch(this, 'props.value', prevProps, prevState, function () {
+        var self = this$1;
+        var ref = self.props;
+        var value = ref.value;
+        if (!self.$f7) { return; }
+        self.setState({
+          currentInputValue: value
+        });
+        self.updateInputOnDidUpdate = true;
+      });
+
+      var self = this;
+      var $listEl = self.$listEl;
+      if (!$listEl || $listEl && $listEl.length === 0) { return; }
+      var isSortable = $listEl.hasClass('sortable');
+
+      if (isSortable !== self.state.isSortable) {
+        self.setState({
+          isSortable: isSortable
+        });
+      }
+
+      var ref = self.props;
+      var validate = ref.validate;
+      var resizable = ref.resizable;
+      var type = ref.type;
+      var f7 = self.$f7;
+      if (!f7) { return; }
+
+      if (self.updateInputOnDidUpdate) {
+        var inputEl = self.refs.inputEl;
+        if (!inputEl) { return; }
+        self.updateInputOnDidUpdate = false;
+
+        if (validate) {
+          self.validateInput(inputEl);
+        }
+
+        if (type === 'textarea' && resizable) {
+          f7.input.resizeTextarea(inputEl);
+        }
+      }
+    };
+
+    F7ListInput.prototype.componentDidMount = function componentDidMount () {
+      var self = this;
+      var el = self.refs.el;
+      if (!el) { return; }
+      self.$f7ready(function (f7) {
+        var ref = self.props;
+        var validate = ref.validate;
+        var resizable = ref.resizable;
+        var value = ref.value;
+        var defaultValue = ref.defaultValue;
+        var type = ref.type;
+        var inputEl = self.refs.inputEl;
+        if (!inputEl) { return; }
+        inputEl.addEventListener('input:notempty', self.onInputNotEmptyBound, false);
+        inputEl.addEventListener('textarea:resze', self.onTextareaResizeBound, false);
+        inputEl.addEventListener('input:empty', self.onInputEmptyBound, false);
+        inputEl.addEventListener('input:clear', self.onInputClearBound, false);
+
+        if ((validate || validate === '') && (typeof value !== 'undefined' && value !== null && value !== '' || typeof defaultValue !== 'undefined' && defaultValue !== null && defaultValue !== '')) {
+          setTimeout(function () {
+            self.validateInput(inputEl);
+          }, 0);
+        }
+
+        if (type === 'textarea' && resizable) {
+          f7.input.resizeTextarea(inputEl);
+        }
+      });
+      self.$listEl = self.$$(el).parents('.list, .list-group').eq(0);
+
+      if (self.$listEl.length) {
+        self.setState({
+          isSortable: self.$listEl.hasClass('sortable')
+        });
+      }
+    };
+
+    prototypeAccessors.slots.get = function () {
+      return __reactComponentSlots(this.props);
+    };
+
+    F7ListInput.prototype.dispatchEvent = function dispatchEvent (events) {
+      var args = [], len = arguments.length - 1;
+      while ( len-- > 0 ) args[ len ] = arguments[ len + 1 ];
+
+      return __reactComponentDispatchEvent.apply(void 0, [ this, events ].concat( args ));
+    };
+
+    prototypeAccessors.refs.get = function () {
+      return this.__reactRefs;
+    };
+
+    prototypeAccessors.refs.set = function (refs) {};
+
+    Object.defineProperties( F7ListInput.prototype, prototypeAccessors );
+
+    return F7ListInput;
+  }(React.Component));
+
+  __reactComponentSetProps(F7ListInput, Object.assign({
+    id: [String, Number],
+    style: Object,
+    className: String,
+    sortable: Boolean,
+    media: String,
+    input: {
+      type: Boolean,
+      default: true
+    },
+    type: {
+      type: String,
+      default: 'text'
+    },
+    name: String,
+    value: [String, Number, Array],
+    defaultValue: [String, Number, Array],
+    readonly: Boolean,
+    required: Boolean,
+    disabled: Boolean,
+    placeholder: String,
+    inputId: [String, Number],
+    size: [String, Number],
+    accept: [String, Number],
+    autocomplete: [String],
+    autocorrect: [String],
+    autocapitalize: [String],
+    spellcheck: [String],
+    autofocus: Boolean,
+    autosave: String,
+    max: [String, Number],
+    min: [String, Number],
+    step: [String, Number],
+    maxlength: [String, Number],
+    minlength: [String, Number],
+    multiple: Boolean,
+    inputStyle: Object,
+    pattern: String,
+    validate: [Boolean, String],
+    tabindex: [String, Number],
+    resizable: Boolean,
+    clearButton: Boolean,
+    noFormStoreData: Boolean,
+    noStoreData: Boolean,
+    ignoreStoreData: Boolean,
+    errorMessage: String,
+    errorMessageForce: Boolean,
+    info: String,
+    label: [String, Number],
+    inlineLabel: Boolean,
+    floatingLabel: Boolean
+  }, Mixins.colorProps));
+
+  F7ListInput.displayName = 'f7-list-input';
+
   var F7ListItemCell = (function (superclass) {
     function F7ListItemCell(props, context) {
       superclass.call(this, props, context);
@@ -4345,7 +4844,8 @@
           hasInputInfo: false,
           hasInputErrorMessage: false,
           hasInputValue: false,
-          hasInputFocused: false
+          hasInputFocused: false,
+          hasInputInvalid: false
         };
       })();
 
@@ -4478,6 +4978,7 @@
       var inlineLabel = props.inlineLabel;
       var itemInputWithInfo = props.itemInputWithInfo;
       var hasInputFocused = self.state.hasInputFocused;
+      var hasInputInvalid = self.state.hasInputInvalid;
       var hasInputValue = self.state.hasInputValue;
       var hasInput = itemInput || self.state.hasInput;
       var hasInlineLabel = inlineLabel || self.state.hasInlineLabel;
@@ -4672,7 +5173,7 @@
         'inline-label': hasInlineLabel,
         'item-input-with-info': hasInputInfo,
         'item-input-with-error-message': hasInputErrorMessage,
-        'item-input-invalid': hasInputErrorMessage,
+        'item-input-invalid': hasInputInvalid,
         'item-input-with-value': hasInputValue,
         'item-input-focused': hasInputFocused
       }, Mixins.colorClasses(props));
@@ -4713,6 +5214,7 @@
       var hasInput = $inputWrapEl.length > 0;
       var hasInputInfo = $inputWrapEl.children('.item-input-info').length > 0;
       var hasInputErrorMessage = $inputWrapEl.children('.item-input-error-message').length > 0;
+      var hasInputInvalid = $inputWrapEl.children('.input-invalid').length > 0;
 
       if (hasInlineLabel !== self.state.hasInlineLabel) {
         self.setState({
@@ -4737,6 +5239,12 @@
           hasInputErrorMessage: hasInputErrorMessage
         });
       }
+
+      if (hasInputInvalid !== self.state.hasInputInvalid) {
+        self.setState({
+          hasInputInvalid: hasInputInvalid
+        });
+      }
     };
 
     F7ListItemContent.prototype.componentDidMount = function componentDidMount () {
@@ -4758,6 +5266,7 @@
       var hasInput = $inputWrapEl.length > 0;
       var hasInputInfo = $inputWrapEl.children('.item-input-info').length > 0;
       var hasInputErrorMessage = $inputWrapEl.children('.item-input-error-message').length > 0;
+      var hasInputInvalid = $inputWrapEl.children('.input-invalid').length > 0;
 
       if (hasInput) {
         el.addEventListener('focus', self.onFocusBound, true);
@@ -4787,6 +5296,12 @@
       if (!self.hasInputErrorMessageSet && hasInputErrorMessage !== self.state.hasInputErrorMessage) {
         self.setState({
           hasInputErrorMessage: hasInputErrorMessage
+        });
+      }
+
+      if (!self.hasInputInvalidSet && hasInputInvalid !== self.state.hasInputInvalid) {
+        self.setState({
+          hasInputInvalid: hasInputInvalid
         });
       }
     };
@@ -5444,7 +5959,7 @@
           }
         }
 
-        if (!tag && 'react' === 'react' || tag && !(tag === 'li' || tag === 'F7ListItem' || tag === 'F7ListButton' || tag.indexOf('list-item') >= 0 || tag.indexOf('list-button') >= 0 || tag.indexOf('f7-list-item') >= 0 || tag.indexOf('f7-list-button') >= 0)) {
+        if (!tag && 'react' === 'react' || tag && !(tag === 'li' || tag === 'F7ListItem' || tag === 'F7ListButton' || tag === 'F7ListInput' || tag.indexOf('list-item') >= 0 || tag.indexOf('list-button') >= 0 || tag.indexOf('list-input') >= 0 || tag.indexOf('f7-list-item') >= 0 || tag.indexOf('f7-list-button') >= 0 || tag.indexOf('f7-list-input') >= 0)) {
           if (wasUlChild) { rootChildrenAfterList.push(child); }else { rootChildrenBeforeList.push(child); }
         } else if (tag) {
           wasUlChild = true;
@@ -7507,7 +8022,8 @@
       this.state = (function () {
         return {
           hasSubnavbar: false,
-          routerClasses: ''
+          routerClass: '',
+          routerForceUnstack: false
         };
       })();
     }
@@ -7548,6 +8064,18 @@
       this.dispatchEvent('page:mounted pageMounted', event, page);
     };
 
+    F7Page.prototype.onPageStack = function onPageStack () {
+      this.setState({
+        routerForceUnstack: false
+      });
+    };
+
+    F7Page.prototype.onPageUnstack = function onPageUnstack () {
+      this.setState({
+        routerForceUnstack: true
+      });
+    };
+
     F7Page.prototype.onPageInit = function onPageInit (event) {
       var page = event.detail;
       var ref = this.props;
@@ -7575,13 +8103,13 @@
 
       if (page.from === 'next') {
         this.setState({
-          routerClasses: 'page-next'
+          routerClass: 'page-next'
         });
       }
 
       if (page.from === 'previous') {
         this.setState({
-          routerClasses: 'page-previous'
+          routerClass: 'page-previous'
         });
       }
 
@@ -7598,13 +8126,13 @@
 
       if (page.to === 'next') {
         this.setState({
-          routerClasses: 'page-next'
+          routerClass: 'page-next'
         });
       }
 
       if (page.to === 'previous') {
         this.setState({
-          routerClasses: 'page-previous'
+          routerClass: 'page-previous'
         });
       }
 
@@ -7614,7 +8142,7 @@
     F7Page.prototype.onPageAfterIn = function onPageAfterIn (event) {
       var page = event.detail;
       this.setState({
-        routerClasses: 'page-current'
+        routerClass: 'page-current'
       });
       this.dispatchEvent('page:afterin pageAfterIn', event, page);
     };
@@ -7693,8 +8221,8 @@
       }
 
       var forceSubnavbar = typeof subnavbar === 'undefined' && typeof withSubnavbar === 'undefined' ? hasSubnavbar || this.state.hasSubnavbar : false;
-      var classes = Utils.classNames(className, 'page', this.state.routerClasses, {
-        stacked: stacked,
+      var classes = Utils.classNames(className, 'page', this.state.routerClass, {
+        stacked: stacked && !this.state.routerForceUnstack,
         tabs: tabs,
         'page-with-subnavbar': subnavbar || withSubnavbar || forceSubnavbar,
         'no-navbar': noNavbar,
@@ -7756,6 +8284,8 @@
       el.removeEventListener('page:afterout', self.onPageAfterOut);
       el.removeEventListener('page:afterin', self.onPageAfterIn);
       el.removeEventListener('page:beforeremove', self.onPageBeforeRemove);
+      el.removeEventListener('page:stack', self.onPageStack);
+      el.removeEventListener('page:unstack', self.onPageUnstack);
     };
 
     F7Page.prototype.componentDidMount = function componentDidMount () {
@@ -7778,6 +8308,8 @@
       self.onPageAfterOut = self.onPageAfterOut.bind(self);
       self.onPageAfterIn = self.onPageAfterIn.bind(self);
       self.onPageBeforeRemove = self.onPageBeforeRemove.bind(self);
+      self.onPageStack = self.onPageStack.bind(self);
+      self.onPageUnstack = self.onPageUnstack.bind(self);
 
       if (ptr) {
         el.addEventListener('ptr:pullstart', self.onPtrPullStart);
@@ -7799,6 +8331,8 @@
       el.addEventListener('page:afterout', self.onPageAfterOut);
       el.addEventListener('page:afterin', self.onPageAfterIn);
       el.addEventListener('page:beforeremove', self.onPageBeforeRemove);
+      el.addEventListener('page:stack', self.onPageStack);
+      el.addEventListener('page:unstack', self.onPageUnstack);
     };
 
     prototypeAccessors.slots.get = function () {
@@ -10997,7 +11531,7 @@
   };
 
   /**
-   * Framework7 React 3.4.3
+   * Framework7 React 3.5.0
    * Build full featured iOS & Android apps using Framework7 & React
    * http://framework7.io/react/
    *
@@ -11005,7 +11539,7 @@
    *
    * Released under the MIT License
    *
-   * Released on: October 19, 2018
+   * Released on: October 26, 2018
    */
 
   var Plugin = {
@@ -11051,6 +11585,7 @@
       window.ListButton = F7ListButton;
       window.ListGroup = F7ListGroup;
       window.ListIndex = F7ListIndex;
+      window.ListInput = F7ListInput;
       window.ListItemCell = F7ListItemCell;
       window.ListItemContent = F7ListItemContent;
       window.ListItemRow = F7ListItemRow;