+++ packages/vue/framework7-vue.esm.d.ts
@@ -40,6 +40,7 @@ import f7Link from './components/link';
 import f7ListButton from './components/list-button';
 import f7ListGroup from './components/list-group';
 import f7ListIndex from './components/list-index';
+import f7ListInput from './components/list-input';
 import f7ListItemCell from './components/list-item-cell';
 import f7ListItemContent from './components/list-item-content';
 import f7ListItemRow from './components/list-item-row';
@@ -153,6 +154,7 @@ export {
   f7ListButton,
   f7ListGroup,
   f7ListIndex,
+  f7ListInput,
   f7ListItemCell,
   f7ListItemContent,
   f7ListItemRow,